+++ stream_alert_cli/utils.py
@@ -850,112 +850,10 @@ def _setup_kinesis_subparser(subparser):
 
 
 
-def _setup_athena_subparser(subparser):
-    """Add athena subparser: manage.py athena [subcommand]"""
-    athena_subparsers = subparser.add_subparsers(dest="athena subcommand", required=True)
-
-    _setup_athena_create_table_subparser(athena_subparsers)
-    _setup_athena_rebuild_subparser(athena_subparsers)
-    _setup_athena_drop_all_subparser(athena_subparsers)
 
 
-def _setup_athena_create_table_subparser(subparsers):
-    """Add the athena create-table subparser: manage.py athena create-table [options]"""
-    athena_create_table_parser = _generate_subparser(
-        subparsers,
-        'create-table',
-        description='Create an Athena table',
-        subcommand=True
-    )
-
-    _set_parser_epilog(
-        athena_create_table_parser,
-        epilog=(
-            '''\
-            Examples:
-
-                manage.py athena create-table \\
-                  --bucket s3.bucket.name \\
-                  --table-name my_athena_table
-            '''
-        )
-    )
-
-    _add_default_athena_args(athena_create_table_parser)
-
-    # Validate the provided schema-override options
-    def _validate_override(val):
-        """Make sure the input is in the format column_name=type"""
-        err = ('Invalid override expression [{}]. The proper format is '
-               '"column_name=value_type"').format(val)
-        if '=' not in val:
-            raise athena_create_table_parser.error(err)
-
-        if len(val.split('=')) != 2:
-            raise athena_create_table_parser.error(err)
-
-    athena_create_table_parser.add_argument(
-        '--schema-override',
-        nargs='+',
-        help=(
-            'Value types to override with new types in the log schema. '
-            'The provided input should be space-separated '
-            'directives like "column_name=value_type"'
-        ),
-        action=UniqueSetAction,
-        default=set(),
-        type=_validate_override
-    )
 
 
-def _setup_athena_rebuild_subparser(subparsers):
-    """Add the athena rebuild-partitions subparser: manage.py athena rebuild-partitions [options]"""
-    athena_rebuild_parser = _generate_subparser(
-        subparsers,
-        'rebuild-partitions',
-        description='Rebuild the partitions for an Athena table',
-        subcommand=True
-    )
-
-    _set_parser_epilog(
-        athena_rebuild_parser,
-        epilog=(
-            '''\
-            Examples:
-
-                manage.py athena rebuild-partitions \\
-                  --bucket s3.bucket.name \\
-                  --table-name my_athena_table
-            '''
-        )
-    )
-
-    _add_default_athena_args(athena_rebuild_parser)
-
-
-def _setup_athena_drop_all_subparser(subparsers):
-    """Add the athena drop-all-tables subparser: manage.py athena drop-all-tables"""
-    _generate_subparser(
-        subparsers,
-        'drop-all-tables',
-        description='Drop all tables from an Athena database',
-        subcommand=True
-    )
-
-
-def _add_default_athena_args(athena_parser):
-    """Adds the default required arguments for athena subcommands (bucket and table)"""
-    athena_parser.add_argument(
-        '-b', '--bucket',
-        help='Name of the S3 bucket where log data is located',
-        required=True
-    )
-
-    athena_parser.add_argument(
-        '-t', '--table-name',
-        help='Name of the Athena table to create. This must be a type of log defined in logs.json',
-        required=True
-    )
 
 
 def _setup_threat_intel_subparser(subparser):
